+++ source/Application/Controller/Admin/NavigationTree.php
@@ -27,15 +27,6 @@ class NavigationTree extends Base
     protected $_oInitialDom = null;
 
     /**
-     * Dynamix XML path
-     *
-     * @deprecated 6.6.0 dynpages will be removed on the next major
-     *
-     * @var string
-     */
-    protected $_sDynIncludeUrl = null;
-
-    /**
      * Default EXPATH supported encodings
      *
      * @var array