+++ source/Core/ViewConfig.php
@@ -1040,18 +1040,6 @@ class ViewConfig extends \OxidEsales\Eshop\Core\Base
     }
 
     /**
-     * Returns service url
-     *
-     * @deprecated 6.6.0 dynpages will be removed on the next major
-     *
-     * @return string
-     */
-    public function getServiceUrl()
-    {
-        return $this->getViewConfigParam('sServiceUrl');
-    }
-
-    /**
      * Returns session Remote Access token. Later you can pass the token over rtoken URL param
      * when you want to access the shop, for example, from different client.
      *