+++ CHANGELOG.md
@@ -191,6 +191,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
         - `_getActiveModuleInfo()`
         - `_getDisabledModuleInfo()`
         - `getModuleTranslationPathFinder()`
+- `OxidEsales\EshopCommunity\Core\Utils::getRemoteCachePath()`
+- `OxidEsales\EshopCommunity\Core\ViewConfig::getServiceUrl()`
+- `OxidEsales\EshopCommunity\Application\Controller\Admin\NavigationTree::$_sDynIncludeUrl`
+- Language constants
+    - `DYN_ECONDA_ACTIVE`
+    - `DYN_ECONDA_ATTENTION`
+    - `DYN_ECONDA_COPY_FILE`
+    - `DYNSCREEN_TITLE`
+    - `DYNSCREEN_LIST_SERVICE`
+    - `LOAD_DYN_CONTENT_NOTICE`
+    - `DYNSCREEN_LOCAL_TITLE`
+    - `DYNSCREEN_LOCAL_TEXT`
+    - `mxdynscreenlocal`
 
 ### Fixed
 - Fix not working actions and promotions [#0005526](https://bugs.oxid-esales.com/view.php?id=5526)